Small improvements

+++ Services/Skill/Tree/class.ilSkillTreeDBRepository.php
@@ -135,7 +135,7 @@ class ilSkillTreeDBRepository implements ilSkillTreeRepository
             [$node_id]
         );
         $rec = $db->fetchAssoc($set);
-        return $rec["skl_tree_id"] ?? 0;
+        return (int) ($rec["skl_tree_id"] ?? 0);
     }
 
     public function getTreeForNodeId(int $node_id): ilSkillTree
